requirements.txt
pynput
pystray
Pillow
